requirements.txt
streamlit==1.44.1
docx2txt==0.9
pandas==2.2.3
PyPDF2==3.0.1
pdfplumber==0.11.6
google-generativeai==0.8.5
plotly== 6.0.1
streamlit-extras==0.6.0
python-dotenv==1.1.0
streamlit-pdf-viewer==0.0.22
python-docx==1.1.2
wordcloud==1.9.4
seaborn
tensorflow
transformers
torch
scikit-learn
supabase